+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Version 1.1.0
+
+Changed options management system with *argparse*.
+New optimization system: *"fast"*, *"tight"*, or *"custom"*.
+Manual updated.
+Bug fixes.
+
 # Version 1.0.0
 
 First release of **pyIHM**.